requirements.txt
-e file:///Users/kenneth/Desktop/lab/alchemist
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
annotated-types==0.7.0
anthropic==0.40.0
anyio==4.7.0
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
deprecated==1.2.15
discord-py==2.4.0
distro==1.9.0
dnspython==2.7.0
docstring-parser==0.16
ecdsa==0.19.0
email-validator==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.7
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib-metadata==8.5.0
iniconfig==2.0.0
itsdangerous==2.2.0
jinja2==3.1.4
jiter==0.8.2
mako==1.3.8
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
mirascope==1.13.1
multidict==6.1.0
openai==1.57.2
openpipe==4.40.0
opentelemetry-api==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
propcache==0.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.4
pydantic-core==2.27.2
pydantic-settings==2.7.0
pygments==2.18.0
pynacl==1.5.0
pytest==8.3.4
pytest-asyncio==0.25.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-lilypad==0.0.10
python-multipart==0.0.20
pyyaml==6.0.2
requests==2.32.3
rich==13.9.4
rich-toolkit==0.12.0
rsa==4.9
ruff==0.8.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.36
sqlmodel==0.0.22
starlette==0.41.3
tqdm==4.67.1
typer==0.15.1
typing-extensions==4.12.2
urllib3==2.2.3
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
websockets==14.1
wrapt==1.17.0
yarl==1.18.3
zipp==3.21.0
